+++ NEWS.md
@@ -4,4 +4,7 @@
 
 # getRad 0.1.0
 
-* New function `get_pvol` downloads and reads polar volume data for 6 countries (#1).
+* Initial package development
+* Add `get_pvol` function to download polar volumes for 6 countries
+* Add `get_vpts` function to download vertical profile time series from the [Aloft data repository](https://aloftdata.eu/)
+* Add `aloft_data_coverage` function to fetch an overview table of the files available on the [Aloft data repository](https://aloftdata.eu/)